Переписал алгоритм при помощи поиска в глубину

+++ src/main/java/org/example/Main.java
@@ -5,15 +5,15 @@ import java.util.List;
 public class Main {
     public static void main(String[] args) {
 
-        long start = System.currentTimeMillis();
+        Long start = System.currentTimeMillis();
 
         List<Float[]> data = DataReader.readData(args[0]);
         List<List<Float[]>> groups = GroupDivider.divide(data);
         DataWriter.writeData(groups);
 
-        long finish = System.currentTimeMillis();
+        Long finish = System.currentTimeMillis();
 
-        System.out.println((float) (finish - start) / 1000);
+        System.out.println((finish - start) / 1000);
     }
 
 }
